+++ src/main/java/com/zsubera/jpa/update/PersistenceContextStrategy.java
@@ -1,0 +1,6 @@
+package com.zsubera.jpa.update;
+
+public enum PersistenceContextStrategy {
+    AUTO_CLEAR,
+    DEFER_TO_CALLER
+}